test-requirements.txt
-r requirements.txt

pytest==4.0.2
pytest-flake8==1.0.2
pytest-sanic==0.1.15
